+++ changelog.md
@@ -53,3 +53,9 @@
   1.2.6
 
 - Reset Screenscraper credetials on menu
+
+  1.2.7
+
+- Reset Screenscraper credetials on menu
+- Fix scraping Gamecube games
+- Added config for enabling / disabling Bezels